make r16598 compile on 64-bit too
svn:r16604

+++ src/or/rendcommon.c
@@ -514,7 +514,7 @@ rend_encode_v2_descriptors(smartlist_t *descs_out,
     ipos_base64 = tor_malloc_zero(ipos_len * 2);
     if (base64_encode(ipos_base64, ipos_len * 2, ipos, ipos_len)<0) {
       log_warn(LD_REND, "Could not encode introduction point string to "
-               "base64. length=%d", ipos_len);
+               "base64. length=%d", (int)ipos_len);
       tor_free(ipos_base64);
       tor_free(ipos);
       return -1;
